refactor(esl-event-listener): move `splitEvents` to internal listener files
