+++ src/modules/esl-event-listener/test/subscribe.test.ts
@@ -1,19 +1,19 @@
 import {EventUtils} from '../core';
 
 describe('EventUtils:subscribe tests', () => {
-  test('EventUtils:subscribe successfully subscribe listener by descriptor', () => {
+  test('EventUtils.subscribe successfully subscribes listener by descriptor', () => {
     const $host = document.createElement('div');
     const handle = jest.fn();
     EventUtils.subscribe($host, {event: 'click'}, handle);
     expect(EventUtils.listeners($host).length).toBe(1);
   });
-  test('EventUtils:subscribe successfully subscribe listener by event name', () => {
+  test('EventUtils.subscribe successfully subscribes listener by event name', () => {
     const $host = document.createElement('div');
     const handle = jest.fn();
     EventUtils.subscribe($host, 'click', handle);
     expect(EventUtils.listeners($host).length).toBe(1);
   });
-  test('EventUtils:subscribe successfully subscribe listener by event provider', () => {
+  test('EventUtils.subscribe successfully subscribes listener by event provider', () => {
     const $host = document.createElement('div');
     const provider = jest.fn(function () {
       expect(this).toBe($host);
@@ -25,7 +25,7 @@ describe('EventUtils:subscribe tests', () => {
     expect(provider).toBeCalledWith($host);
   });
 
-  test('EventUtils:subscribe successfully subscribe listeners by string with multiple events', () => {
+  test('EventUtils.subscribe successfully subscribes listeners by string with multiple events', () => {
     const $host = document.createElement('div');
     const handle = jest.fn();
     EventUtils.subscribe($host, 'click keydown', handle);
