+++ src/modules/esl-utils/dom/events/test/misc.test.ts
@@ -5,8 +5,7 @@ import {
   isPassiveByDefault,
   getTouchPoint,
   getOffsetPoint,
-  getCompositeTarget,
-  splitEvents
+  getCompositeTarget
 } from '../misc';
 
 describe('dom/events: misc', () => {
@@ -115,36 +114,5 @@ describe('dom/events: misc', () => {
         y:  (boundingClientRect.top + window.scrollY)
       });
     });
-  });
-
-  describe('splitEvents', () => {
-    test(
-      'Empty string parsed to an empty array without exceptions',
-      () => expect(splitEvents('')).toEqual([])
-    );
-    test(
-      'Single event string parsed to single event',
-      () => expect(splitEvents('abc')).toEqual(['abc'])
-    );
-    test(
-      'Single event string with extra spaces parsed to single event',
-      () => expect(splitEvents('  abc ')).toEqual(['abc'])
-    );
-    test(
-      'Multiple events string parsed to separate events',
-      () => expect(splitEvents('abc de f')).toEqual(['abc', 'de', 'f'])
-    );
-    test(
-      'Multiple events string with extra spaces parsed to separate events',
-      () => expect(splitEvents(' a b')).toEqual(['a', 'b'])
-    );
-    test(
-      'The same events count ones when parsed (two unique terms)',
-      () => expect(splitEvents(' a a b')).toEqual(['a', 'b'])
-    );
-    test(
-      'The same events count ones when parsed (one unique term)',
-      () => expect(splitEvents('a a a a a')).toEqual(['a'])
-    );
   });
 });
